HSW AUTOMATIZACION/requirements.txt
anthropic>=0.25.0
feedparser>=6.0.11
requests>=2.31.0
python-dotenv>=1.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.125.0
gspread>=6.0.2
rich>=13.7.0
click>=8.1.7
python-dateutil>=2.9.0
